+++ include/ExtraPopup.hpp
@@ -108,11 +108,11 @@ protected:
 
         // secret buttons
         if (isDeveloperMode()) {
-            spr = ButtonSprite::create("Create in\neditor", "bigFont.fnt", "GJ_button_01.png", scale1);
-            spr->setScale(scale2);
-            menu->addChildAtPosition(
-                CCMenuItemSpriteExtra::create(spr, this, menu_selector(ExtraOptionsPopup::devCreateInEditor)), 
-                Anchor::TopRight, ccp(50, -20));
+            // spr = ButtonSprite::create("Create in\neditor", "bigFont.fnt", "GJ_button_01.png", scale1);
+            // spr->setScale(scale2);
+            // menu->addChildAtPosition(
+            //     CCMenuItemSpriteExtra::create(spr, this, menu_selector(ExtraOptionsPopup::devCreateInEditor)), 
+            //     Anchor::TopRight, ccp(50, -20));
         }
 
         updateGroupInfoLabel();
@@ -160,16 +160,16 @@ private:
         updateButtons();
     }
 
-    void devCreateInEditor(CCObject*) {
-        if (auto obj = EditorUI::get()->m_selectedObject) {
-            CCPoint end;
-            auto str = m_myGroup->toString(obj->getPosition(), &end);
-            LevelEditorLayer::get()->createObjectsFromString(str, 1, 1);
-        } else {
-            shortAlert("Pivot object not selected");
-        }
-        onClose(nullptr);
-    }
+    // void devCreateInEditor(CCObject*) {
+    //     if (auto obj = EditorUI::get()->m_selectedObject) {
+    //         CCPoint end;
+    //         auto str = m_myGroup->toString(obj->getPosition(), &end);
+    //         LevelEditorLayer::get()->createObjectsFromString(str, 1, 1);
+    //     } else {
+    //         shortAlert("Pivot object not selected");
+    //     }
+    //     onClose(nullptr);
+    // }
 
     void updateButtons() {
         bool active = m_groupCmiInfo.m_groupHasSelectedCmi;