+++ changelog.md
@@ -1,3 +1,9 @@
+# 2.1.0
+- Group pinning
+  - <cp>Pinning</c> groups so that they stay on the screen
+  - Dragging pinned groups across the screen
+  - Faster addition/removal of objects in pinned groups (without having to place the object in the editor)
+
 # 2.0.0 (a complete mod rewrite)
 - Group names
 - <cg>Full-fledged in-game group configuration</c>
@@ -6,7 +12,7 @@
   - Changing button positions within the build tabs
   - Changing button positions within each group
   - Editing group icons
-  - And more ...
+  - You can find more information about it [here](https://github.com/RazoomGD/geode-object-groups/blob/v2/Configuring%20Object%20Groups%20v2.md)
 - Additional <cp>tabs</c>
 - Group search (experimental, disabled by default)
 - Amazing <cl>default group configuration</c>